# static/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Netflix movies and TV shows. API UI developed by swgger",
		"version": "1.0",
		"title": "Netflix titles"

	},
	"host": "127.0.0.1",
	"basePath": "/",
...
	"paths": {
...
		"/update": {
			"post": {
				"tags": ["Titles"],
				"summary": "Add a new movie/tv show to the list",
				"description": "",
				"operationId": "add_record",
				"consumes": ["application/json", "application/xml"],
				"produces": ["application/json", "application/xml"],
				"parameters": [{
					"in": "body",
					"name": "body",
					"description": "movie/tv object that needs to be added to the list",
					"required": true,
					"schema": {
						"$ref": "#/definitions/shows"
					}
				}],
				"responses": {
					"405": {
						"description": "Invalid input"
					},
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/ApiResponse"
						}
					}
				}
			}
		},
		"/put": {

			"put": {
				"tags": ["Titles"],
				"summary": "Update an existing movie/tv show",
				"description": "",
				"operationId": "update_record",
				"consumes": ["application/json", "application/xml"],
				"produces": ["application/json", "application/xml"],
				"parameters": [{
					"in": "body",
					"name": "body",
					"description": "movie/tv show object that needs to be updated to the list",
					"required": true,
					"schema": {
						"$ref": "#/definitions/shows"
					}
				}]

			}
		},
		"/delete": {
			"delete": {
				"tags": ["Titles"],
				"summary": "Remove movie/tv show by id",
				"description": "",
				"operationId": "delete_record",
				"produces": [""],
				"parameters": [{
					"show_id": "id",
					"description": "the show ID",
					"required": true,
					"type": "string"

				}]


			}
		}
	},
...
}